requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
autocommand==2.2.2
babel==2.16.0
backoff==2.2.1
backports.tarfile==1.2.0
bangla==0.0.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.35.12
botocore==1.35.12
cachetools==5.5.0
catalogue==2.0.10

click==8.1.7
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
debugpy==1.8.5
decorator==5.1.1
deepspeed==0.8.3
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
einops==0.8.0
elevenlabs==1.11.0
encodec==0.1.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.3
fastjsonschema==2.20.0
ffmpeg==1.4
ffmpy==0.4.0
fire==0.6.0
Flask==3.0.3
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
g2pkk==0.1.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
gradio==5.4.0
gradio_client==1.4.2
grpcio==1.66.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.4.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jamo==0.4.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.1
jieba==0.42.1
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonlines==1.2.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.0
lightning==2.4.0
lightning-cloud==0.5.70
lightning-utilities==0.11.7
lightning_sdk==0.1.15
litdata==0.2.19
litserve==0.2.2
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl-service==2.4.0
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.0.5
murmurhash==1.0.10
mutagen==1.47.0
narwhals==1.10.0
nbclient==0.10.0
nbconvert==5.3.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
ninja==1.11.1.1
nltk==3.9.1
notebook_shim==0.2.4
num2words==0.5.13
numba==0.60.0
numpy==1.22.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai==1.52.2
ordered-set==4.1.0
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
playsound==1.3.0
pooch==1.8.2
preshed==3.0.9
progressbar==2.5
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.23.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pynndescent==0.5.13
pyparsing==3.1.4
pypinyin==0.53.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-json-logger==2.0.7
python-multipart==0.0.12
pytorch-lightning==2.4.0
pytz==2024.1
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests-oauthlib==2.0.0
resampy==0.4.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rotary-embedding-torch==0.3.6
rpds-py==0.20.0
rsa==4.9
ruff==0.7.1
s3transfer==0.10.2
safehttpx==0.1.1
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
simple-term-menu==1.6.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.0
SudachiDict-core==20241021
SudachiPy==0.6.8
sympy==1.13.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
testpath==0.6.0
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.20.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.5.0
torchaudio==2.5.0
torchmetrics==1.3.1
torchvision==0.17.1+cu121
tornado==4.2
tortoise-tts @ git+https://github.com/neonbjb/tortoise-tts@2d6d3868f99d2fb97429741aba7d3a4314ae04aa
tqdm==4.66.5
trainer==0.0.36
traitlets==5.14.3
transformers==4.46.0
triton==3.1.0
-e git+https://github.com/coqui-ai/TTS@dbf1a08a0d4e47fdad6172e433eeb34bc6b13b4e#egg=TTS
typeguard==4.3.0
typer==0.12.5
types-python-dateutil==2.9.0.20240821
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
yarl==1.9.11
zipp==3.19.2
